Fixes #18, Fixes #19 Added Support for class level annotation Removed restriction to be used only on JDK1.7 and above. Compiled with JDK-6 Refactored package name
