+++ .codecov.yml
@@ -1,18 +1,7 @@
-# Team Yaml
-coverage:
-  round: down
-  precision: 5
-
 # Repository Yaml
 coverage:
   round: up
   range: 0..10
-
-# Used in Codecov after updating
-coverage:
-  round: up
-  range: 0..10
-  precision: 5
 
 codecov:
   branch: master
