+++ src/test/java/com/venomvendor/gson/NullDefenseTypeAdapterFactoryTest.java
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package com.venomvendor.library.gson;
+package com.venomvendor.gson;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.TypeAdapterFactory;
 import com.google.gson.reflect.TypeToken;
-import com.venomvendor.library.gson.annotation.MandatoryTest;
-import com.venomvendor.library.gson.model.ChildTest;
-import com.venomvendor.library.gson.model.ParentTest;
+import com.venomvendor.gson.annotation.MandatoryTest;
+import com.venomvendor.gson.model.ChildTest;
+import com.venomvendor.gson.model.ParentTest;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Tag;
